Split gizmos.py into submodules, final task of #218
